feat: create a new types module for hover components and update imports
