+++ src/renderer/components/states/hover/utils/CountdownEdit.tsx
@@ -29,8 +29,7 @@ import { useTranslation } from 'react-i18next';
 import useIslandStore from '../../../../store/slices';
 import { SvgIcon } from '../../../../utils/SvgIcon';
 import { ToolButtons } from './ToolButtons';
-
-type TimerState = 'idle' | 'running' | 'paused';
+import type { TimerState } from '../types';
 
 function padZero(value: number): string {
   return value < 10 ? `0${value}` : `${value}`;
